+++ includes/ray.h
@@ -25,7 +25,7 @@ typedef struct s_ray
 
 typedef struct s_intersection
 {
-	float	t;
+	double	t;
 	t_obj	*obj;
 }	t_intersection;
 
@@ -36,10 +36,10 @@ typedef struct s_intersections
 }	t_intersections;
 
 t_ray			ray(t_tuple origin, t_tuple direction);
-t_ray			ray_for_pixel(t_camera cam, float x, float y);
-t_tuple			position(t_ray ray, float t);
+t_ray			ray_for_pixel(t_camera cam, double x, double y);
+t_tuple			position(t_ray ray, double t);
 void			intersect(t_ray r, t_obj *o, t_intersections *xs);
-t_intersection	intersection(float t, t_obj *obj);
+t_intersection	intersection(double t, t_obj *obj);
 t_intersection	*hit(t_intersections *inters);
 
 t_ray			ray_transform(t_ray r, t_mtx4 m);
