+++ src/maths/fcolors.c
@@ -27,7 +27,7 @@ t_fcolor	color_sub(t_fcolor c1, t_fcolor c2)
 	return ((t_fcolor){c1.r - c2.r, c1.g - c2.g, c1.b - c2.b});
 }
 
-t_fcolor	color_scalar(t_fcolor c1, float k)
+t_fcolor	color_scalar(t_fcolor c1, double k)
 {
 	return ((t_fcolor){c1.r * k, c1.g * k, c1.b * k});
 }
@@ -40,7 +40,7 @@ t_fcolor	color_mul(t_fcolor c1, t_fcolor c2)
 uint32_t	fcolor_to_uint(t_fcolor col)
 {
 	uint32_t	res;
-	float		buf;
+	double		buf;
 
 	res = 0;
 	buf = col.r * 256 - 0.5;
