+++ src/maths/lighting/phong.c
@@ -20,7 +20,7 @@ t_fcolor	phong(t_material m, t_light light, t_pre_compute *pc)
 	t_tuple const	lightv = tp_normalize(tp_sub(light.pos, pc->pos));
 	t_fcolor		diffuse;
 	t_fcolor		specular;
-	float			reflect_dot_eye;
+	double			reflect_dot_eye;
 
 	diffuse = fcolor(0, 0, 0);
 	specular = fcolor(0, 0, 0);
