+++ src/maths/ray/intersections.c
@@ -17,7 +17,7 @@
 #include "rt_maths.h"
 #include "lib/libft/src/vector/vector.h"
 
-t_intersection	intersection(float t, t_obj *obj)
+t_intersection	intersection(double t, t_obj *obj)
 {
 	return ((t_intersection){t, obj});
 }
@@ -25,9 +25,9 @@ t_intersection	intersection(float t, t_obj *obj)
 void	intersect_calc(t_ray r, t_obj *o, t_intersections *xs)
 {
 	t_tuple const	d = tp_sub(r.origin, point(0, 0, 0));
-	float const		b = 2 * tp_dot(r.dir, d);
-	float const		a = tp_dot(r.dir, r.dir);
-	float const		dis = b * b - 4 * a * (tp_dot(d, d) - 1);
+	double const		b = 2 * tp_dot(r.dir, d);
+	double const		a = tp_dot(r.dir, r.dir);
+	double const		dis = b * b - 4 * a * (tp_dot(d, d) - 1);
 
 	if (dis >= 0)
 	{
